Consolidate NuGetUtils and NuGetUtilities classes
